Cambios en seeder que elimina imagens con agregacion de prompt

+++ database/seeders/ClearImagesSeeder.php
@@ -21,21 +21,34 @@ class ClearImagesSeeder extends Seeder
         // Imagen que siempre quieres conservar
         $exceptFileName = 'no-photo.png';
 
-        // Lista de nombres de imágenes asociadas a productos
-        $productImages = Producto::pluck('image')->toArray(); // Asume que la columna se llama 'image'
-        $productImages[] = $exceptFileName; // También agregas la imagen por defecto
+
 
         // Obtiene todos los archivos en el directorio
         $filePaths = $fs->files(storage_path('app/public/images'));
 
-        foreach ($filePaths as $filePath) {
-            $fileName = basename($filePath);
-
-            // Solo elimina si el archivo NO está en la lista de imágenes asociadas a productos
-            if (!in_array($fileName, $productImages)) {
-                $fs->delete($filePath);
+        if ($this->command->confirm('¿Desea eliminar todas las imagenes del STORAGE?', true)) {
+            foreach ($filePaths as $filePath) {
+                $fileName = basename($filePath);
+                // Verifica si el archivo es el que deseas conservar
+                if ($fileName !== $exceptFileName) {
+                    $fs->delete($filePath);
+                }
             }
+            echo "Archivos eliminados exitosamente!\n";
         }
-        echo "Archivos eliminados exitosamente!\n";
+        if ($this->command->confirm('¿Desea eliminar solo las imagenes no asociadas a productos?', true)) {
+            // Lista de nombres de imágenes asociadas a productos
+            $productImages = Producto::pluck('image')->toArray(); // Asume que la columna se llama 'image'
+            $productImages[] = $exceptFileName; // También agregas la imagen por defecto
+            foreach ($filePaths as $filePath) {
+                $fileName = basename($filePath);
+                // Solo elimina si el archivo NO está en la lista de imágenes asociadas a productos
+                if (!in_array($fileName, $productImages)) {
+                    $fs->delete($filePath);
+                }
+            }
+            echo "Imagenes no asociadas eliminadas!\n";
+        }
+
     }
 }
